+++ src/Fantomas.Core.Tests/CodeFormatterTests.fs
@@ -1,7 +1,9 @@
 module Fantomas.Core.Tests.CodeFormatterTests
 
 open NUnit.Framework
+open Fantomas.FCS.Text
 open Fantomas.Core
+open Fantomas.Core.SyntaxOak
 open Fantomas.Core.Tests.TestHelpers
 
 [<Test>]
@@ -39,7 +41,7 @@ let main _ =
     |> ignore
 
 [<Test>]
-let ``trivia is parsed for Oak`` () =
+let ``trivia is transformed to Oak`` () =
     let oak =
         CodeFormatter.ParseOakAsync(false, "let a = 0\n // foo")
         |> Async.RunSynchronously
@@ -67,3 +69,61 @@ let ``parsed oak can be formatted back to source`` () =
         |> Async.RunSynchronously
 
     Assert.AreEqual(source, formatted)
+
+[<Test>]
+let ``transform parsedInput to Oak`` () =
+    let source =
+        """
+module A
+
+#if DEBUG
+let b = 0
+#endif
+"""
+
+    let ast, _ =
+        Fantomas.FCS.Parse.parseFile false (SourceText.ofString source) [ "DEBUG" ]
+
+    let oak = CodeFormatter.TransformAST(ast, source)
+
+    match oak.ModulesOrNamespaces.[0].Declarations.[0] with
+    | ModuleDecl.TopLevelBinding _ -> Assert.Pass()
+    | _ -> Assert.Fail()
+
+[<Test>]
+let ``transform parsedInput created with additional defines to Oak`` () =
+    let source =
+        """
+module A
+
+#if DEBUG
+let b = 0
+#endif
+"""
+
+    let ast, _ =
+        Fantomas.FCS.Parse.parseFile false (SourceText.ofString source) [ "DEBUG"; "FOO"; "BAR" ]
+
+    let oak = CodeFormatter.TransformAST ast
+
+    match oak.ModulesOrNamespaces.[0].Declarations.[0] with
+    | ModuleDecl.TopLevelBinding _ -> Assert.Pass()
+    | _ -> Assert.Fail()
+
+[<Test>]
+let ``transform parsedInput contains trivia in Oak`` () =
+    let source =
+        """
+module A
+
+// foo
+let b = 0
+"""
+
+    let ast, _ = Fantomas.FCS.Parse.parseFile false (SourceText.ofString source) []
+
+    let oak = CodeFormatter.TransformAST(ast, source)
+
+    match oak.ModulesOrNamespaces.[0].Declarations.[0] with
+    | ModuleDecl.TopLevelBinding node -> Assert.True node.HasContentBefore
+    | _ -> Assert.Fail()